Adding swal message before delete variant

+++ src/varient/index.js
@@ -6,6 +6,7 @@ import AddCircleIcon from "@mui/icons-material/AddCircle";
 import { collection, getDocs, deleteDoc, doc } from "firebase/firestore";
 import { db } from "../firebase-config";
 import VarientList from "./components/VarientList";
+import Swal from "sweetalert2";
 
 function Varient() {
   const [addVarient, setVarient] = useState(false);
@@ -28,8 +29,26 @@ function Varient() {
   };
 
   const handleDelete = async (id) => {
-    await deleteDoc(doc(ref, id));
-    setVarientData(varientData.filter((row) => row.id !== id));
+    const result = await Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this! Kindly remove products from variant before delete it !!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Yes, delete it!',
+      cancelButtonText: 'Cancel',
+      confirmButtonColor: "#1976d2",
+      cancelButtonColor: "#ff5722",
+    });
+
+    if (result.isConfirmed) {
+      await deleteDoc(doc(ref, id));
+      setVarientData(varientData.filter((row) => row.id !== id));
+      Swal.fire(
+        'Deleted!',
+        'Your variant has been deleted.',
+        'success'
+      );
+    }
   };
 
   const handleSearch = (event) => {
